fix(cute_dsl): make the optional-dependency guard independent of cutlass (#4753)
Move availability probes to a cutlass-free module so import flashinfer
does not hard-require nvidia-cutlass-dsl. Paths that do not exist on
0.6.18 (kda_prefill_cute, SM120 KDA prefill runtime, attn_scores) are
omitted.
